added aggregates without groupBy and made THAT code work

+++ thefacetakt/src/main/java/ru/fizteh/fivt/students/thefacetakt/collectionsql/impl/aggregates/CountVisitor.java
@@ -5,17 +5,17 @@ import ru.fizteh.fivt.students.thefacetakt.collectionsql.AggregatorVisitor;
 
 import java.util.function.Function;
 
-public class CountVisitor<T> implements AggregatorVisitor<T, Integer> {
+public class CountVisitor<T> implements AggregatorVisitor<T, Long> {
     private Function<T, ?> function;
-    private Integer count;
+    private Long count;
 
     public CountVisitor(Function<T, ?> newFunction) {
-        count = 0;
+        count = 0L;
         function = newFunction;
     }
 
     @Override
-    public Integer result() {
+    public Long result() {
         return count;
     }
